subtopics entity moved to domain.layer
