+++ Servers/database/db.ts
@@ -11,7 +11,7 @@ import { ProjectScopeModel } from "../domain.layer/models/projectScope/projectSc
 import { ProjectsMembersModel } from "../domain.layer/models/projectsMembers/projectsMembers.model";
 import { QuestionModel } from "../domain.layer/models/question/question.model";
 import { SubcontrolModel } from "../domain.layer/models/subcontrol/subcontrol.model";
-import { SubtopicModel } from "../models/subtopic.model";
+import { SubtopicModel } from "../domain.layer/models/subtopic/subtopic.model";
 import { TopicModel } from "../models/topic.model";
 import { UserModel } from "../domain.layer/models/user/user.model";
 import { VendorModel } from "../models/vendor.model";
