+++ Servers/domain.layer/models/question/question.model.ts
@@ -5,7 +5,7 @@ import {
   Model,
   Table,
 } from "sequelize-typescript";
-import { SubtopicModel } from "../../../models/subtopic.model";
+import { SubtopicModel } from "../subtopic/subtopic.model";
 
 /*
