+++ Servers/domain.layer/models/subtopic/subtopic.mock.data.ts
@@ -1,199 +1,201 @@
-import { Subtopic } from "../models/subtopic.model";
+import { Subtopic } from "./subtopic.model";
 
-export const subtopics = (
-  topics: number[]
-): Subtopic[] => {
+export const subtopics = (topics: number[]): Subtopic[] => {
   return [
     {
       topic_id: topics[0],
       title: "General",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[0],
       title: "Technology details",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[1],
       title: "Transparency and provision of information to deployers",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[1],
       title: "Responsibilities along the AI value chain",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[2],
       title: "Responsibilities along the AI value chain",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[2],
       title: "Fundamental rights impact assessments for high-risk AI systems",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[3],
       title: "AI model capability assessment",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[4],
       title: "AI model capability assessment",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[5],
       title: "User notification of AI system use",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[6],
       title: "Oversight documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[6],
       title: "Human intervention mechanisms",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[7],
       title: "System validation and reliability documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[7],
       title: "AI system change documentation",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[8],
       title: "EU database registration",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[9],
-      title: "Post-market monitoring by providers and post-market monitoring plan for high-risk AI systems",
-      order_no: 1
+      title:
+        "Post-market monitoring by providers and post-market monitoring plan for high-risk AI systems",
+      order_no: 1,
     },
     {
       topic_id: topics[10],
       title: "Bias and fairness evaluation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[11],
       title: "System information documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[12],
-      title: "Transparency obligations for providers and users of certain AI systems",
-      order_no: 1
+      title:
+        "Transparency obligations for providers and users of certain AI systems",
+      order_no: 1,
     },
     {
       topic_id: topics[13],
       title: "Environmental impact",
-      order_no: 1
+      order_no: 1,
     },
 
     {
       topic_id: topics[14],
       title: "General",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[14],
       title: "Technology details",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[15],
       title: "Transparency and provision of information to deployers",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[15],
       title: "Responsibilities along the AI value chain",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[16],
       title: "Responsibilities along the AI value chain",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[16],
       title: "Fundamental rights impact assessments for high-risk AI systems",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[17],
       title: "AI model capability assessment",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[18],
       title: "AI model capability assessment",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[19],
       title: "User notification of AI system use",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[20],
       title: "Oversight documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[20],
       title: "Human intervention mechanisms",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[21],
       title: "System validation and reliability documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[21],
       title: "AI system change documentation",
-      order_no: 2
+      order_no: 2,
     },
     {
       topic_id: topics[22],
       title: "EU database registration",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[23],
-      title: "Post-market monitoring by providers and post-market monitoring plan for high-risk AI systems",
-      order_no: 1
+      title:
+        "Post-market monitoring by providers and post-market monitoring plan for high-risk AI systems",
+      order_no: 1,
     },
     {
       topic_id: topics[24],
       title: "Bias and fairness evaluation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[25],
       title: "System information documentation",
-      order_no: 1
+      order_no: 1,
     },
     {
       topic_id: topics[26],
-      title: "Transparency obligations for providers and users of certain AI systems",
-      order_no: 1
+      title:
+        "Transparency obligations for providers and users of certain AI systems",
+      order_no: 1,
     },
     {
       topic_id: topics[27],
       title: "Environmental impact",
-      order_no: 1
+      order_no: 1,
     },
-  ]
+  ];
 };
