+++ Servers/utils/user.utils.ts
@@ -29,7 +29,7 @@ import { FileModel } from "../models/file.model";
 import { ControlCategoryModel } from "../domain.layer/models/controlCategory/controlCategory.model";
 import { AssessmentModel } from "../domain.layer/models/assessment/assessment.model";
 import { TopicModel } from "../models/topic.model";
-import { SubtopicModel } from "../models/subtopic.model";
+import { SubtopicModel } from "../domain.layer/models/subtopic/subtopic.model";
 import { QuestionModel } from "../domain.layer/models/question/question.model";
 import {
   createOrganizationQuery,
